show-utils {S4Vectors}	R Documentation

Display utilities

Description

Low-level utilities that control display of vector-like objects.

Usage

get_showHeadLines()
set_showHeadLines(n=5)

get_showTailLines()
set_showTailLines(n=5)

Arguments

n
A non-negative integer that controls the number of vector elements to display.

Details

For the sake of keeping display compact, the show() methods for Vector derivatives only display 5 head and 5 tail vector elements.

However, the number of head and tail elements to display can be changed by setting global options showHeadLines and showTailLines to the desired values.

get_showHeadLines(), set_showHeadLines(), get_showTailLines(), and set_showTailLines() are convenience functions for getting/setting these global options.

Value

get_showHeadLines() and get_showTailLines() return the current showHeadLines and showTailLines values.

set_showHeadLines() and set_showTailLines() return the showHeadLines and showTailLines values before the change, invisibly.

See Also

options in base R.

Vector objects.

Examples

library(IRanges)
ir <- IRanges(start=11:45, width=10)
ir  # displays 5 head and 5 tail ranges

set_showHeadLines(18)
ir  # displays 18 head ranges

set_showHeadLines()  # back to default

[Package S4Vectors version 0.48.0 Index]