zip-methods {S4Vectors}	R Documentation

Convert between parallel vectors and lists

Description

The zipup and zipdown functions convert between two parallel vectors and a list of doublets (elements of length 2). The metaphor, borrowed from Python's zip, is that of a zipper. The zipup function interleaves the elements of the parallel vectors into a list of doublets. The inverse operation is zipdown, which returns a Pairs object.

Usage


zipup(x, y, ...)
zipdown(x, ...)

Arguments

x,y
For zipup, any vector-like object. For zipdown, a doublet list.

...
Arguments passed to methods.

Value

For zipup, a list-like object, where every element is of length 2. For zipdown, a Pairs object.

See Also

Pairs objects.

Examples


z <- zipup(1:10, Rle(1L, 10))
pairs <- zipdown(z)

[Package S4Vectors version 0.20.1 Index]