bindROWS {S4Vectors}	R Documentation

Combine objects by ROWS or COLS

Description

bindROWS and bindCOLS are low-level generic functions defined in the S4Vectors package for binding objects along their 1st or 2nd dimension. They are the workhorses behind higher-level operations like c(), rbind(), or cbind() on most vector-like or rectangular objects defined in Bioconductor. They are not intended to be used directly by the end user.

Usage

bindROWS(x, objects=list(), use.names=TRUE, ignore.mcols=FALSE, check=TRUE)
bindCOLS(x, objects=list(), use.names=TRUE, ignore.mcols=FALSE, check=TRUE)

Arguments

x
An S4 object.

objects
A list of S4 objects to bind to x. They should typically (but not necessarily) have the same class as x.

use.names
Should the names on the input objects be propagated? By default they are.

ignore.mcols
Should the metadata columns on the input objects be ignored? By defaut they are not (i.e. they are propagated).

check
Should the result object be validated before being returned to the user? By default it is.

Value

An object of the same class as x.

Author(s)

Hervé Pagès

See Also

The NROW and NCOL generic functions defined in the BiocGenerics package.

[Package S4Vectors version 0.28.1 Index]